Created and pushed configWizard.py and helga.py
helga.py is the main code, this is the control center from which all the other scripts and stuff can be controlled and called.

configWizard.py is called when first starting the bot, and each time after, it confirms the settings the bot will use and saves them to a Config.ini file upon changing them.